requirements.txt
pandas
numpy
scikit-learn
sentence-transformers
faiss-cpu
tqdm
rank-bm25
lightgbm
joblib
matplotlib
flask
pyarrow
fastparquet
